Disable SubStorage features of XStorage until https://gforge.statsbiblioteket.dk/tracker/index.php?aid=1186 is resolved.
Make XStorage respect the contract the ConfigurationStorage.get should not throw NPE when the key is unknwon, merely return null.

Make XStorage use an internal XProperties instead of directly inheriting from XProperties - this had a lot of tricky problems when trying to serialize the properties in the constructor of XStorage.

Add a log4j configuration to the tests in Common module

Remove obsolete/buggy testcase from ConfigurationStorageTestCase