The overall summary of this commit is that H2Storage passes the unit tests now...
I added a new method DatabaseStorage.isIntegrityConstraintViolation(SQLException e) that is used to check if an SQLException represents a ICV. The problem is that H2 does not use the standard JDBC ICVException for this, but just some custom error codes.

Fix StorageTest to put databases inside a root 'test_db' directory. This fixes the cleanup of the H2 databases which do not use the same layout as the Derby one...

Rename H2 jar from datestamped version to 1.0.79 version stamped name