Create new class ManipulatorFactory to help instantiate IndexManipulators
Move the CONF_MANIPULATOR_CLASS  property def. from IndexControllerImpl to the IndexManipulator interface

Make IndexControllerImpl use the new ManipulatorFactory for instantiating its manipulators

Introduce a RemoteManipulator interface and a RMIManipulatorProxy that can be used to expose a manipulator over RMI. The intention is to be able to remote control commit() and consolidate() via shell commands and WorkflowSteps in the Workflow module.

Introduce properly defined default property valies in RMIStorageProxy

Make StorageFactory use the static getClass() method in Configuration to avoid RMI mess-ups of class paths.

Don't throw RemoteExceptions in StorageFactory when we should really be throwing ConfigurationExceptions...