New feature: A Status web service, exposing a single method (for now) fullStatus which collects status from the searcher, storage, and suggest services and formats an XML reply with the whole deal.
What's missing to make this complete is a display xslt for the returned XML as well as a method variant returning RSS or Atom.

Longer term we want to implement special features in each service to convey a more detailed status message. In particular when the index was last updated.